final/builds/tests/requirements.txt
pytest
filelock
pymysql
sqlalchemy
pytest-xdist
pytest-check
allure-pytest
requests
selenium
webdriver_manager
bs4
